add app permission check

+++ darwin-web/src/main/java/xin/manong/darwin/web/controller/AppController.java
@@ -12,6 +12,7 @@ import xin.manong.darwin.service.iface.AppService;
 import xin.manong.darwin.web.convert.Converter;
 import xin.manong.darwin.web.request.AppRequest;
 import xin.manong.darwin.web.request.AppUpdateRequest;
+import xin.manong.darwin.web.service.AppPermissionService;
 
 import javax.annotation.Resource;
 import javax.ws.rs.*;
@@ -33,6 +34,8 @@ public class AppController {
 
     @Resource
     protected AppService appService;
+    @Resource
+    protected AppPermissionService appPermissionService;
 
     /**
      * 根据应用名搜索应用
@@ -136,6 +139,7 @@ public class AppController {
             logger.error("app is not found for id[{}]", request.id);
             throw new NotFoundException(String.format("应用[%d]不存在", request.id));
         }
+        appPermissionService.checkAppPermission(request.id);
         App app = Converter.convert(request);
         return appService.update(app);
     }
@@ -159,6 +163,7 @@ public class AppController {
             logger.error("app is not found for id[{}]", id);
             throw new NotFoundException(String.format("应用[%d]不存在", id));
         }
+        appPermissionService.checkAppPermission(id);
         return appService.delete(id);
     }
 }
